created home page, changed files location
