+++ src/screens/categories/Combined Print and E-Book Fiction.js
@@ -7,12 +7,12 @@ import {
   Text,
 } from "react-native";
 import tw from "tailwind-react-native-classnames";
+import BookDetail from "../../components/modals/BookDetail";
 
 import { useSelector, useDispatch } from "react-redux";
-import { booksSelector, fetchBooks } from "../slices/books";
-import BookDetail from "../components/modals/BookDetail";
+import { booksSelector, fetchBooks } from "../../slices/books";
 
-export default function YoungAdultHardcover() {
+export default function CombinedPrintAndEBookFiction() {
   const dispatch = useDispatch();
   const { loading, hasErrors, books } = useSelector(booksSelector);
   const [isDetail, setIsDetail] = useState(false);
@@ -23,14 +23,14 @@ export default function YoungAdultHardcover() {
   const handleDetailConfirm = () => {
     setIsDetail(false);
   };
-
   useEffect(() => {
     dispatch(fetchBooks());
   }, []);
-  console.log("books", books[10]);
+
+  console.log("books", books[0]);
   return (
     <ScrollView>
-      {books[10]?.books.map((book) => (
+      {books[0]?.books.map((book) => (
         <TouchableOpacity
           key={book.book_uri}
           onPress={() => {
