+++ src/screens/categories/Combined Print and E-Book Nonfiction.js
@@ -7,12 +7,11 @@ import {
   Text,
 } from "react-native";
 import tw from "tailwind-react-native-classnames";
-
 import { useSelector, useDispatch } from "react-redux";
-import { booksSelector, fetchBooks } from "../slices/books";
-import BookDetail from "../components/modals/BookDetail";
+import { booksSelector, fetchBooks } from "../../slices/books";
+import BookDetail from "../../components/modals/BookDetail";
 
-export default function Science() {
+export default function CombinedPrintAndEBookNonfiction() {
   const dispatch = useDispatch();
   const { loading, hasErrors, books } = useSelector(booksSelector);
   const [isDetail, setIsDetail] = useState(false);
@@ -20,17 +19,17 @@ export default function Science() {
   const [author, setAuthor] = useState("");
   const [description, setDescription] = useState(null);
 
-  const handleDetailConfirm = () => {
+  function handleDetailConfirm() {
     setIsDetail(false);
-  };
+  }
 
   useEffect(() => {
     dispatch(fetchBooks());
   }, []);
-  console.log("books", books[14]);
+  console.log("books", books[1]);
   return (
     <ScrollView>
-      {books[14]?.books.map((book) => (
+      {books[1]?.books.map((book) => (
         <TouchableOpacity
           key={book.book_uri}
           onPress={() => {
