+++ src/screens/categories/Hardcover Fiction.js
@@ -9,10 +9,10 @@ import {
 import tw from "tailwind-react-native-classnames";
 
 import { useSelector, useDispatch } from "react-redux";
-import { booksSelector, fetchBooks } from "../slices/books";
-import BookDetail from "../components/modals/BookDetail";
+import { booksSelector, fetchBooks } from "../../slices/books";
+import BookDetail from "../../components/modals/BookDetail";
 
-export default function TradeFictionPaperback() {
+export default function HardcoverFiction() {
   const dispatch = useDispatch();
   const { loading, hasErrors, books } = useSelector(booksSelector);
   const [isDetail, setIsDetail] = useState(false);
@@ -27,10 +27,10 @@ export default function TradeFictionPaperback() {
   useEffect(() => {
     dispatch(fetchBooks());
   }, []);
-  console.log("books", books[4]);
+  console.log("books", books[2]);
   return (
     <ScrollView>
-      {books[4]?.books.map((book) => (
+      {books[2]?.books.map((book) => (
         <TouchableOpacity
           key={book.book_uri}
           onPress={() => {
